{{ cookiecutter.project_name }}/composition_repository/fobi/requirements.txt
django-fobi==0.19.8
